==> src/02.7_speciesMeta.R <==
source(file.path(Sys.getenv("CODEBASE"),"DNAmeth500species/src/00.0_init.R"))

wd=file.path(analysis_dir,"02_vizStats/02.7_speciesMeta")
dir.create(wd,recursive=TRUE)
setwd(wd)

#remove the unconverted
stats_annot=stats_annot[!grepl("_uc$",Sample_Name)&species_check!="fail"&!grepl("Tumour|Cellline",Tissue)] 

stats_annot[,Tissue:=factor(Tissue,levels=core_vert),]
stats_annot_spec=stats_annot[,.(cphph=mean(CpG_meth_cphph),cphpg=mean(CpG_meth_cphpg),cpg=mean(CpG_meth),pdr=mean(PDR_mean)),by=c("species","color_class")]


stats_annot[color_class == "Jawless_vertebrate",group:="Inv."]
stats_annot[color_class == "Jawless_vertebrate",color_class:="Invertebrata"]
#plot stats for non-cpg methylation and pdr

#average per class
ggplot(stats_annot,aes(y=CpG_meth_cphpg,x=group,fill=color_class))+
  geom_boxplot(outlier.shape = 21)+
  scale_fill_manual(values = class_colors)+
  geom_point(data = stats_annot[species=="JL", ], aes(y=CpG_meth_cphpg,x=group), shape =21, fill = class_colors["Jawless_vertebrate"])+
  stat_summary(fun.data = give.n,fun.args = c(y=-0.1), geom = "text",size=2)+
  xlab("")+ylab("")

ggplot(stats_annot,aes(y=PDR_mean,x=group,fill=color_class))+
  geom_boxplot(outlier.shape = 21)+
 geom_point(data = stats_annot[species=="JL", ], aes(y=CpG_meth_cphpg,x=group), shape =21, fill = class_colors["Jawless_vertebrate"])+
  scale_fill_manual(values = class_colors)+
  stat_summary(fun.data = give.n,fun.args = c(y=-0.1), geom = "text",size=2)+
  xlab("")+ylab("")


#average per tissue
pdf("tissue.pdf",10,4.5)
ggplot(stats_annot[Tissue%in%core_vert],aes(y=CpG_meth_cphph,x=Tissue,fill=color_class))+
  geom_boxplot(size=0.2,outlier.size = 0.5,outlier.shape = 21)+
  scale_fill_manual(values = class_colors)+
  stat_summary(fun.data = give.n,fun.args = c(y=-0.1), geom = "text",size=2)+
  scale_x_discrete(labels=class_short)+xlab("")+ylab("% DNA methylation at non-CpGs")+
  facet_wrap(~color_class,nrow=2)+rotate_labels()

ggplot(stats_annot[Tissue%in%core_vert],aes(y=PDR_mean,x=Tissue,fill=color_class))+
  geom_boxplot(size=0.2,outlier.size = 0.5,outlier.shape = 21)+
  scale_fill_manual(values = class_colors)+
  stat_summary(fun.data = give.n,fun.args = c(y=0), geom = "text",size=2)+
  scale_x_discrete(labels=class_short)+xlab("")+ylab("Average PDR")+
  facet_wrap(~color_class,nrow=2)+rotate_labels()+ylim(0,1)

ggplot(stats_annot[Tissue%in%core_vert],aes(y=CpG_meth,x=Tissue,fill=color_class))+
  geom_boxplot(size=0.2,outlier.size = 0.5,outlier.shape = 21)+
  scale_fill_manual(values = class_colors)+
  stat_summary(fun.data = give.n,fun.args = c(y=-0.1), geom = "text",size=2)+
  scale_x_discrete(labels=class_short)+xlab("")+ylab("% DNA methylation at CpGs")+
  facet_wrap(~color_class,nrow=2)+rotate_labels()

dev.off()

#focus on brain non-cpg methylation
test=function(a,b){
  if (sum(!is.na(a))<2|sum(!is.na(b))<2){return(NA)}
  wt=wilcox.test(a,b,paired = T,alternative="greater")
  return(wt$p.value)
}

brain_non_cpg=stats_annot[species%in%stats_annot[Tissue=="Brain"]$species,.(brain=mean(CpG_meth_cphph[Tissue=="Brain"],na.rm=TRUE),non_brain=mean(CpG_meth_cphph[Tissue!="Brain"],na.rm=TRUE)),by=c("color_class","species")]
brain_non_cpg[,N:=.N,by=c("color_class")]
#brain_non_cpg=brain_non_cpg[N>1]
brain_non_cpg[,color_class2:=ifelse(color_class%in%c("Mammalia","Aves"),as.character(color_class),"Others"),]

pvals=brain_non_cpg[,.(p=test(brain, non_brain)),by=c("color_class2")]

brain_non_cpg[,.(median(brain/non_brain)),by=c("color_class2")]
#color_class2       V1
#1:        other 1.029880
#2:         Aves 1.591792
#3:     Mammalia 1.718065


pdf("brain_non_CpGmeth.pdf",5,3.5)
ggplot(brain_non_cpg,aes(y=log2(brain/non_brain),x=color_class2,color=color_class))+
  geom_boxplot(color="black")+
  geom_text(aes(label=species), position=position_jitter(width=0.2))+
  geom_text(data=pvals,color="black",aes(label=paste0("p=",signif(p,3))),y=-1,size=3)+
  scale_color_manual(values = class_colors)+
  geom_hline(yintercept = 0, lty=11)+
  ylab("Non-CpG methylation\nlog2(brain/non-brain)")+
  scale_x_discrete(labels=class_short,drop=TRUE)
dev.off()

#PDR vs. mean methylation
pdf("methVsPDR.pdf",5,3)
ggplot(stats_annot,aes(y=PDR_mean,x=CpG_meth,color=color_class))+
  geom_point(alpha=0.2)+
    geom_point(data = stats_annot[species=="JL", ], aes(y=PDR_mean,x=CpG_meth), color = class_colors["Jawless_vertebrate"], alpha = 0.2)+
  scale_color_manual(values = class_colors)+
  geom_smooth(color="black")+
  xlab("% DNA methylation at CpGs")+ylab("Average PDR")+
  xlim(c(0,100))
dev.off()

#PDR vs. prefragmentation
pdf("prefragVsPDR.pdf",5,3)
ggplot(stats_annot,aes(y=PDR_mean,x=others,color=color_class))+
  geom_point(alpha=0.2)+
  
  scale_color_manual(values = class_colors)+
  geom_smooth(color="black")+
  xlab("DNA prefragmentation (%)")+ylab("Average PDR")
dev.off()



#median PDR per class vs. median DNA meth per class

class_med=stats_annot[,.(median_PDR=median(PDR_mean),median_CpG_meth=median(CpG_meth),N=.N),by="color_class"]

cor=class_med[,cor.test(median_CpG_meth,median_PDR)]

pdf("median_PDRvsMeth.pdf",5,3)
ggplot(class_med, aes(x=median_CpG_meth, y=median_PDR,color=color_class))+
  geom_smooth(method="lm",color="black", se=F)+
  geom_point(aes(size=N))+
  geom_text(aes(label=N),color="black")+
  annotate("text",hjust=0,label=paste0("p=",signif(cor$p.value,3),"\nr=",signif(cor$estimate,3)),x=30,y=0.3,color="black")+
  scale_color_manual(values = class_colors)+
  scale_size_continuous(range = c(6,17))+xlim(0,100)+ylim(0,NA)+
  ylab("Median PDR")+xlab("Median CpG methylation (%)")
dev.off()
  



#look at all measures in Lamprey
JL_stats=melt(stats_annot[,c("CpG_meth","CpG_meth_cphph","PDR_mean",'Sample_Name',"color_class","species"),],id.vars = c('Sample_Name',"color_class","species"))
JL_stats[,plot_group:=ifelse(species=="JL","Lamprey",ifelse(color_class=="Invertebrata","Invertebrata","Vertebrata")),]

#cap outliers for better plotting
JL_stats[,value:=ifelse(value>quantile(value,0.99),quantile(value,0.99),value),by=c("variable")]

pdf("Lamprey.pdf",5.5,4)
ggplot(JL_stats,aes(x=value,fill=plot_group,col=plot_group))+geom_density(lwd=1,alpha=0.3)+
  geom_point(data=JL_stats[species=="JL"],y=0, size=3)+facet_wrap(~variable,scale="free",ncol=2)
dev.off()

#look at bimodal CpG meth in invertebrates
stats_inv=stats_annot[color_class=="Invertebrata"]
stats_inv[,med_meth:=median(CpG_meth),by="species"]
stats_inv[,scientific_name:=factor(scientific_name,levels=unique(scientific_name[order(med_meth)])),]
median_meth=stats_annot[,.(median=median(CpG_meth)),by="color_class"]

pdf("invertebrates.pdf",9,4)
ggplot(stats_inv,aes(x=scientific_name,y=CpG_meth))+
  geom_hline(data=median_meth,aes(yintercept = median,color=color_class),lty=11,lwd=0.2)+
  geom_point(shape=21)+
  scale_color_manual(values = class_colors)+rotate_labels()
dev.off()

#investigate extremes in non-cpg methylation
cphph_extremes=table(stats_annot[CpG_meth_cphph>2.5]$species)

cphph_extremes_spec=names(cphph_extremes[cphph_extremes>1])

ggplot(stats_annot[species%in%cphph_extremes_spec],aes(y=CpG_meth_cphph,x=species,color=color_class))+
  geom_point()+
  geom_hline(yintercept = 2.5)+
  scale_color_manual(values = class_colors)+
  stat_summary(fun.data = give.n,fun.args = c(y=-0.1), geom = "text",size=4)+
  scale_x_discrete(labels=class_short)+xlab("")+ylab("")


#check for non-cpg meth biases due to technical reasons (none visible)
ggplot(stats_annot,aes(y=k1_unmeth,x=CpG_meth_cphph,color=color_class,shape=species%in%cphph_extremes_spec))+
  geom_point()+
  scale_color_manual(values = class_colors)


ggplot(stats_annot,aes(y=cont_rat,x=CpG_meth_cphph,color=color_class,shape=species%in%cphph_extremes_spec))+
  geom_point()+
  scale_color_manual(values = class_colors)



#check for PDR biases due to technical reasons (none visible)
ggplot(stats_annot,aes(y=PDR_mean,x=CpG_meth,color=color_class))+
  geom_point()+
  scale_color_manual(values = class_colors)

ggplot(stats_annot,aes(y=PDR_mean,x=others,color=color_class))+
  geom_point()+
  scale_color_manual(values = class_colors)

ggplot(stats_annot,aes(y=PDR_mean,x=cont_rat,color=color_class))+
  geom_point()+
  scale_color_manual(values = class_colors)


##association with cancer risk
rel=stats_annot[!is.na(`Adult weight (g)`)&!is.na(`Maximum longevity (yrs)`)]

rel[,.N,by=c('abbreviation_sp','ncbi_name','color_class')]
rel[,length(unique(ncbi_name)),by=c('color_class')]

id_vars=c('abbreviation_sp','ncbi_name','color_class','Adult weight (g)','Maximum longevity (yrs)')

rel_red=rel[,.(CpG_meth=mean(CpG_meth),CpG_meth_cphph=mean(CpG_meth_cphph),PDR_mean=mean(PDR_mean)),
    by=id_vars]

rel_long=melt(rel_red,id.vars=id_vars)

rel_long[,MLTAW:=log(`Maximum longevity (yrs)`^6 * `Adult weight (g)`),]

rel_long=rel_long[color_class%in%c('Mammalia','Aves','Actinopteri')]


#get correlations
get_cor=function(x,y){
  r=cor.test(x,y)
  return(list('cor'=r$estimate,'p'=r$p.value,x=min(x),y=max(y)))
}

cor_w=rel_long[,get_cor(`Adult weight (g)`,value),by=c('color_class','variable')]
cor_l=rel_long[,get_cor(`Maximum longevity (yrs)`,value),by=c('color_class','variable')]
cor_c=rel_long[,get_cor(MLTAW,value),by=c('color_class','variable')]


cor_w[,p_adj:=p.adjust(p,method = 'fdr'),]
cor_l[,p_adj:=p.adjust(p,method = 'fdr'),]
cor_c[,p_adj:=p.adjust(p,method = 'fdr'),]


#plot relationships
ggplot(rel_long,aes(x=`Adult weight (g)`,y=value))+
  geom_text(data=cor_w,vjust=1,hjust=0,aes(x=x,y=y,label=paste0('p=',signif(p_adj,3),'\nr=',signif(cor,3))))+
  geom_point()+geom_smooth(method = 'lm')+
  facet_wrap(color_class~variable, scale='free',nrow=3)

ggplot(rel_long,aes(x=`Maximum longevity (yrs)`,y=value))+
  geom_text(data=cor_l,vjust=1,hjust=0,aes(x=x,y=y,label=paste0('p=',signif(p_adj,3),'\nr=',signif(cor,3))))+
  geom_point()+geom_smooth(method = 'lm')+
  facet_wrap(color_class~variable, scale='free',nrow=3)

pdf("MLTAW.pdf",7,6)
ggplot(rel_long,aes(x=MLTAW,y=value))+ geom_text(data=cor_c,vjust=1,hjust=0,aes(x=x,y=y,label=paste0('p=',signif(p_adj,3),'\nr=',signif(cor,3))),size=3)+
  geom_point()+geom_smooth(method = 'lm')+
  facet_wrap(variable~color_class, scale='free',nrow=3)
dev.off()


#association with cancer risk phyloglm
library(phylolm)

group="Aves"
#group="Mammalia"

tree=ape::read.tree(paste0(analysis_dir,"/01_basicStats/01.6_ITOL/tree_species_as_tips.phy"))

data=data.frame(rel_long[variable=="CpG_meth"&color_class==group&!duplicated(abbreviation_sp),c("abbreviation_sp","value","MLTAW"),],row.names = "abbreviation_sp")
colnames(data)[1]="CpG_meth"
head(data)

tree=drop.tip(tree,tip = tree$tip.label[!tree$tip.label%in%row.names(data)])
str(tree)

#check distribution. Here looks like gaussian
hist(data$MLTAW)

fit = compar.gee(MLTAW~CpG_meth,data=data,family="gaussian",phy=tree)
fit_nophy= glm(MLTAW~CpG_meth,data=data,family="gaussian")

fit
summary(fit_nophy)
